pyproject.toml
[project]
name = "potify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "spotipy>=2.25.1",
    "typer>=0.16.0",
]
